% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addBangForBuck}
\alias{addBangForBuck}
\title{addBangForBuck}
\usage{
addBangForBuck(pathsToTarget, STUDY, dtd)
}
\arguments{
\item{pathsToTarget}{the set of pathways to be targeted}

\item{STUDY}{the study object}

\item{dtd}{drug target data table as returned by getDrugTargetData()}
}
\value{
dtd table with "number of paths containing gene" column added
}
\description{
adds bang for buck analysis which shows how many paths each genes exists in.
}

