% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BangForBuck}
\alias{BangForBuck}
\title{Get pathway overlaps.}
\usage{
BangForBuck(darkPaths, path_detail)
}
\arguments{
\item{darkPaths}{The paths to search for overlaps.}

\item{path_detail}{A Path_Detail object.}
}
\value{
A data frame with columns "Number of paths" and "Path names", and a row for each gene in all the submitted pathways.
}
\description{
Determine the set of overlapping genes and number of pathway overlaps for a set of pathways.
}

