% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_paint.R
\name{.loadGraphitePaths}
\alias{.loadGraphitePaths}
\title{load graphite-supplied pathways}
\usage{
.loadGraphitePaths(stud)
}
\arguments{
\item{stud}{The \code{Study} object}
}
\value{
The pathway repository selected.
}
\description{
Internal function used to ensure graphite supplied network diagrams are loaded and made available for cell network display.
}
