% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getBasicSettings}
\alias{getBasicSettings}
\title{Get settings object pre-loaded for example analysis}
\usage{
getBasicSettings()
}
\value{
A \code{Settings}
}
\description{
Returns set of setting for analysis of example somatic mutation data, example drug screen data, overlap analysis, and path analysis.
}

