% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGenesFromPaths}
\alias{getGenesFromPaths}
\title{Get gene identifiers from a list of pathways.}
\usage{
getGenesFromPaths(pids, STUDY)
}
\arguments{
\item{pids}{Character vector of cellular pathway identifiers.}

\item{STUDY}{A study object, a Path_Detail object or a bipartate graph in a "matrix" class R object.}
}
\value{
Character vector of gene identifiers.
}
\description{
Takes a list of path identifiers and returns a list of identifiers for the genes in the submitted pathways.
}

