% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/path_paint.R
\name{getHumanPathIds}
\alias{getHumanPathIds}
\title{Pull out human path ids from the reactome.db annotation package}
\usage{
getHumanPathIds()
}
\value{
\code{list}, names are path names, values are db ids.
}
\description{
Get human path ids from the reactome.db annotation package
}

