\name{getLoadedPathSummaryRunner}
\alias{getLoadedPathSummaryRunner}
\title{Get \code{path_summary_runner} object.}
\usage{
getLoadedPathSummaryRunner()
}
\value{
\code{path_summary_runner} object loaded with the data that
is provided for path significance test functions.
}
\description{
The \code{path_summary_runner} object returned by this
function is one of the two passed by summaryTable to the
path significance tests.
}
\seealso{
\code{\link{summaryTable}}
}

