% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPathIdsToTarget}
\alias{getPathIdsToTarget}
\title{getPathIdsToTarget}
\usage{
getPathIdsToTarget(STUDY, pathsToSearch = NULL, verbose = TRUE)
}
\arguments{
\item{STUDY}{the study object}

\item{pathsToSearch}{the set of pathways to search for targets. If this is provided, function passes this input as the output}
}
\value{
Vector of path ids/path names
}
\description{
getPathIdsToTarget
}

