\name{getPathsFromStudy}
\alias{getPathsFromStudy}
\title{getPathsFromStudy}
\usage{
getPathsFromStudy(studyObject)
}
\arguments{
  \item{studyObject}{An object of class Study}
}
\value{
a path matrix: representation of a collection of pathway
gene memberships as a bipartate graph, with columns
representing genes and rows representing pathways.
}
\description{
Retreives biological pathway bipartate graph from study
object
}

