% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biopaxToCytoscape.R
\name{getReactomeIds}
\alias{getReactomeIds}
\title{Get reactome database ids, give reactome path ids.}
\usage{
getReactomeIds(pnames)
}
\arguments{
\item{pnames}{A vector of pathway names}
}
\value{
A two column data frame; columns pid (path name) and dbid (database id). Row names are set as path names.
}
\description{
Get reactome database ids, give reactome path ids. Used in forming URL strings to download biopax files.
}

