\name{getTestStudyObject}
\alias{getTestStudyObject}
\title{Get a Study object containing moc data.}
\usage{
getTestStudyObject()
}
\value{
\code{Study} object
}
\description{
Returns a \code{Study} object loaded with moc somatic
mutation data which indicates the Reactome pathway,
Abacavir metabolism, is enriched in somatic mutation.
}

