% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biopaxToCytoscape.R
\name{nodeNamesFromLabels}
\alias{nodeNamesFromLabels}
\title{nodeNamesFromLabels}
\usage{
nodeNamesFromLabels(w, nodeLabels)
}
\arguments{
\item{w}{a cytoscape connection object}

\item{nodeLabels}{a vector of node labels}
}
\value{
a vector of node names corresponding to the supplied node labels
}
\description{
nodeNamesFromLabels
}

