% Generated by roxygen2 (4.0.2): do not edit by hand
\name{patientSummaryToNozzleReport}
\alias{patientSummaryToNozzleReport}
\title{Make a Nozzle report from an individual patient's results}
\usage{
patientSummaryToNozzleReport(patSum, rootReportName, resSetName)
}
\arguments{
\item{patSum}{The list element, including item name of the patient's summary (note this must include the patient name, thus what is given by \code{list[1]} would be correct, not what is subscripted like \code{list[[1]]})}

\item{rootReportName}{The file path to the root nozzle report ex: "./output/study_mutSigAnalysisOnly/results/somatic_mutation_aberration_summary.nozzleReport"}

\item{resSetName}{}
}
\value{
The path to the file that the patient summary was saved to.
}
\description{
Make a Nozzle report from an individual patient's results
}

