\name{printProgramState}
\alias{printProgramState}
\title{printProgramState}
\usage{
printProgramState(stud)
}
\arguments{
  \item{stud}{A Study object from which the program state
  is to be determined.}
}
\description{
Displays to standard out general information about the
state of the program, including pathways, settings and data
currently loaed.
}

