% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/save_and_load_data.R
\name{saveStudy}
\alias{saveStudy}
\title{saveStudy}
\usage{
saveStudy(study, path = "./output")
}
\arguments{
\item{study}{A Study object to be saved.}

\item{path}{The file path where the study folder should be saved.}
}
\description{
Function for saving study object to a human-readable format.
}

