\name{setBorderColors}
\alias{setBorderColors}
\title{setBorderColors}
\usage{
setBorderColors(w, color_choices, targeted = NULL, active,
  default.color = "#a0a0a0")
}
\arguments{
  \item{w}{cytoscape connection object,}

  \item{color_choices}{vector of color 4 choices, the
  default, the targeted & not active nodes, the targeted
  and active nodes and the color to use if there is only
  coverage data}

  \item{targeted}{vector of targeted genes (all strings)}

  \item{active}{vector of active genes (all strings)}

  \item{default.color}{the default}
}
\description{
setBorderColors
}

