\name{armDescriptionList}
\alias{armDescriptionList}
\title{Get all arm descriptions}
\usage{
armDescriptionList(study)
}
\arguments{
  \item{study}{A \code{Study} object}
}
\value{
\code{vector} of study arm descriptions.
}
\description{
Retreives arm descriptions from the arms slot of a Study
object. This thus contains descriptions of all the arms
loaded and ready for use.
}

