% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHumanPathIds}
\alias{getHumanPathIds}
\title{Pull out human path ids from the reactome.db annotation package}
\usage{
getHumanPathIds()
}
\value{
\code{list}, names are path names, values are db ids.
}
\description{
Get human path ids from the reactome.db annotation package
}

