% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drugDbInterface.R
\name{getPathIdsToTarget}
\alias{getPathIdsToTarget}
\title{getPathIdsToTarget}
\usage{
getPathIdsToTarget(STUDY, pathsToSearch = NULL, verbose = TRUE)
}
\arguments{
\item{STUDY}{the study object}

\item{pathsToSearch}{the set of pathways to search for targets. If this is provided, function passes this input as the output}

\item{verbose}{Logical flag indicating if diagnostic information should be printed to the screen.}
}
\value{
Vector of path ids/path names
}
\description{
getPathIdsToTarget
}

