\name{getPathIdsToTarget}
\alias{getPathIdsToTarget}
\title{getPathIdsToTarget}
\usage{
getPathIdsToTarget(STUDY, pathsToSearch = NULL)
}
\arguments{
  \item{pathsToSearch}{the set of pathways to search for
  targets. If this is provided, function passes this input
  as the output}

  \item{STUDY}{the study object}
}
\value{
Vector of path ids/path names
}
\description{
getPathIdsToTarget
}

