\name{getPathwaysRecords}
\alias{getPathwaysRecords}
\title{getPathwaysRecords}
\usage{
getPathwaysRecords(pwrecord.fileName = NULL)
}
\arguments{
  \item{pwrecord.fileName}{the file name for the pathway
  records file.}
}
\value{
data frame with columns: dbID, path_name and download_date
}
\description{
Opens or creates record of downlaoded biopax pathways
}

