% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nodeNamesFromLabels}
\alias{nodeNamesFromLabels}
\title{nodeNamesFromLabels}
\usage{
nodeNamesFromLabels(w, nodeLabels)
}
\arguments{
\item{w}{a cytoscape connection object}

\item{nodeLabels}{a vector of node labels}
}
\value{
a vector of node names corresponding to the supplied node labels
}
\description{
nodeNamesFromLabels
}

