% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_functions.R
\name{openPGM}
\alias{openPGM}
\title{openPGM}
\usage{
openPGM(fname = NULL)
}
\arguments{
\item{fname}{The file path to the patient gene matrix file.}
}
\value{
The patient gene matrix (logic matrix with gene identifiers given as row names and patient identifiers gien as column names)
}
\description{
Opens a patient gene matrix file.
}

