\name{saveDataWorkUpNotes}
\alias{saveDataWorkUpNotes}
\title{saveDataWorkUpNotes}
\usage{
saveDataWorkUpNotes(set, fpath)
}
\arguments{
  \item{set}{the data work up notes}

  \item{fpath}{the file path that any tables should be
  saved to.}
}
\value{
data.frame with two columns: the notes and the values of
the notes.
}
\description{
Transforms the data work up notes, from a list, as the are
produced by the arms,
}

