\name{saveStudy}
\alias{saveStudy}
\title{saveStudy}
\usage{
saveStudy(study, path = "./output/")
}
\arguments{
  \item{study}{A Study object to be saved.}

  \item{path}{The file path where the study folder should
  be saved.}
}
\description{
Function for saving study object to a human-readable
format.
}

