\name{SaveToHTML}
\alias{SaveToHTML}
\title{Save results to an HTML page.}
\usage{
SaveToHTML(study)
}
\arguments{
  \item{study}{A \code{Study} object with results sets from
  one or more study arm.}
}
\description{
Writes data from the \code{results} slot of a \code{Study}
object to an HTML document.
}

