% Generated by roxygen2 (4.0.2): do not edit by hand
\name{whichPaths}
\alias{whichPaths}
\title{whichPaths}
\usage{
whichPaths(pathList, geneList, STUDY)
}
\arguments{
\item{pathList}{the set of pathways to be selected from}

\item{geneList}{the list of genes}

\item{STUDY}{the study object}
}
\value{
A vector of path IDs
}
\description{
Indicates the set of pathways a particular set of genes belongs to
}

