\name{whichPaths}
\alias{whichPaths}
\title{whichPaths}
\usage{
whichPaths(pathList, geneList, STUDY)
}
\arguments{
  \item{pathList}{the set of pathways to be selected from}

  \item{STUDY}{the study object}

  \item{geneList}{the list of genes}
}
\value{
A vector of path IDs
}
\description{
Indicates the set of pathways a particular set of genes
belongs to
}

